chore: AuthOptions not assignable to type never
