+++ bciers/apps/registration1/app/components/operations/Operations.tsx
@@ -2,7 +2,7 @@ import { GridRowsProp } from "@mui/x-data-grid";
 
 import { actionHandler } from "@/app/utils/actions";
 import { getServerSession } from "next-auth";
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import { authOptions } from "@/app/utils/auth/authOptions";
 import {
   OperationRow,
   OperationsSearchParams,
