+++ bciers/apps/registration1/app/layout.tsx
@@ -7,14 +7,14 @@ You should not manually add <head> tags such as <title> and <meta> to root layou
 
 // eslint-disable-next-line import/extensions
 import "@bciers/styles/globals.css";
-import SessionProvider from "../components/auth/SessionProvider";
+import SessionProvider from "@/app/components/auth/SessionProvider";
 import type { Metadata, Viewport } from "next";
 import Footer from "@bciers/components/layout/Footer";
 import Header from "@bciers/components/layout/Header";
 // 🏷 import {named} can be significantly slower than import default
 import Box from "@mui/material/Box";
 import { getServerSession } from "next-auth";
-import { authOptions } from "./api/auth/[...nextauth]/route";
+import { authOptions } from "@/app/utils/auth/authOptions";
 import { PublicEnvScript } from "next-runtime-env";
 import ThemeProvider from "@mui/material/styles/ThemeProvider";
 import { theme } from "@bciers/components";
